refactor into single redux class
