+++ typings/rematch/index.d.ts
@@ -126,3 +126,9 @@ export interface ConfigRedux {
   createStore?: Redux.StoreCreator,
   devtoolOptions?: Object,
 }
+
+declare global {
+  interface Window {
+    __REDUX_DEVTOOLS_EXTENSION_COMPOSE__?: any,
+  }
+}
